Refactored to support delete.
The new algorithm loads all the existing buildings in the list. This
means we know what buildings in firestore no longer exist in the new
database. In addition, the new rule for updating a building's revision
is whether any data already in firestore is different for a matching
building name.