Improve db update logging for import
I tried out running this import script on a copy of the existing data.
The most obvious issue was that revision was being set to NaN
when buildings didn't already have a revision key. It's also useful
to have easier-to-parse logging that can tell you what fields were
different and led to updating the entry in the database.